minzufs/requirements.txt
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-formtools==2.3
django-import-export==2.5.0
django-reversion==3.0.9
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
Keras==2.3.1
Markdown==3.1.1
numpy==1.19.0
opencv-python==4.3.0.36
Pillow==8.3.0
PyJWT==1.7.1
PyMySQL==0.10.0
scipy==1.4.1
tensorboard==2.1.1
tensorflow==2.1.0
django==2.2.5



